requirements.txt
fastapi==0.109.2
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
pydantic==2.6.1
pydantic-settings==2.1.0
google-generativeai==0.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

python-multipart==0.0.9
httpx==0.26.0
python-dotenv==1.0.1
alembic==1.13.1
aiosqlite
web3
eth-account
